fix(catalog providers): separation and capabilities
- Treats all providers derived from the models.dev catalogue as
  independent providers
    - Rather than treating their models as if they were provided by
      OpenCode
- Moves generic catalogue-derived provider code out of opencode.rs into
  a dedicated catalog.rs module
- Fixes capabilities (e.g., vision) for OpenCode and other catalogue
  providers by deriving the data from the models.dev catalogue
- Separates OpenCode Go and Zen, displaying them as separate providers
- Sorts the list of providers alphabetically
