+++ maki-providers/src/providers/mod.rs
@@ -12,6 +12,7 @@ use tracing::debug;
 use crate::AgentError;
 
 pub(crate) mod anthropic;
+pub(crate) mod catalog;
 pub(crate) mod copilot;
 pub mod custom;
 pub(crate) mod deepseek;
